% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qset_mbqs.R
\docType{data}
\name{qset_mbqs}
\alias{qset_mbqs}
\title{MBQS Q-set criteria scores and derived scales}
\format{MBQS dataset with 90 rows (items) and 4 columns.
 \describe{
   \item{item}{Item number from 1 to 90.}
   \item{qset}{mbqs. Maternal Behaviour Q-set (version 3.1) (Pederson et al., 1999).}
   \item{sens_s}{Sensibility criterion scores
   (retrieved from Pederson et al., 1999).}
   \item{description}{Item description
   (retrieved from Pederson et al., 1999).}}}
\usage{
qset_mbqs
}
\description{
A data frame containing the Maternal Behaviour Q-set (mbqs; version 3.1). 
In this dataset, column names ending in "_c" refer to criteria 
sorts' scores, while column names ending in "_s" refer to
scales' scores.
}
\references{
Pederson, D. R., Moran, G., & Bento, S. (1999). Maternal Behaviour Q-sort
 (version 3.1). London, ON: Psychology Department, Western University.
}
\keyword{datasets}
