% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsets.R
\docType{data}
\name{qsets}
\alias{qsets}
\title{Q-sets criteria scores and derived scales}
\format{\enumerate{
\item \strong{aqs} AQS dataset with 90 rows (items) and 9 columns
\describe{
  \item{item}{item number from 1 to 90}
  \item{qset}{aqs. Attachment Q-set (version 3.0) (Waters, 1995)}
  \item{sec_c}{security criterion scores (retrieved from Waters et al., 1995)}
  \item{dep_c}{dependency criterion scores (retrieved from Waters et al., 1995)}
  \item{pederson_s}{Pederson & Moran's (1995) scales (retrieved from Waters et al., 1995)}
  \item{pederson_s_inv}{items to be inverted in Pederson and Moran scales}
  \item{posada_s}{Posada et al.'s (1995) scales (retrieved from Waters et al., 1995)}
  \item{posada_s_inv}{items to be inverted in Posada et al.'s scales}
  \item{description}{item description (retrieved from Waters, 1995)}
}

\item \strong{ccq} CCQ dataset with 100 rows (items) and 6 columns
\describe{
  \item{item}{item number from 1 to 100}
  \item{qset}{ccq. California Child Q-set (Block & Block, 1969)}
  \item{scomp_c}{social competence criterion scores (retrieved from Waters et al., 1985)}
  \item{sest_c}{self-esteem criterion scores (retrieved from Waters et al., 1985)}
  \item{sdes_c}{social desirability criterion scores (retrieved from Waters et al., 1985)}
  \item{description}{item description (retrieved from Waters et al., 1995)}
}

\item \strong{mbqs} MBQS dataset with 90 rows (items) and 4 columns
\describe{
  \item{item}{item number from 1 to 90}
  \item{qset}{mbqs. Maternal Behaviour Q-set (version 3.1) (Pederson et al., 1999)}
  \item{sens_s}{sensibility criterion scores (retrieved from Pederson et al., 1999)}
  \item{description}{item description (retrieved from Waters et al., 1985)}
}

\item \strong{pq} PQ dataset with 72 rows (items) and 6 columns
\describe{
  \item{item}{item number from 1 to 72}
  \item{qset}{pq. Preschool Q-set (Baumrind, 1968 revised by Wanda Bronson)}
  \item{scomp_c}{social competence criterion scores (retrieved from Waters et al., 1985)}
  \item{sest_c}{self-esteem criterion scores (retrieved from Waters et al., 1985)}
  \item{sdes_c}{social desirability criterion scores (retrieved from Waters et al., 1985)}
  \item{description}{item description (retrieved from Waters et al., 1995)}
}
}}
\usage{
qsets
}
\description{
A list containing four different datasets: Attachment Q-set (AQS; version
3.0), California Child Q-set (CCQ); Maternal Behaviour Q-set (MBQS; version
3.1); and Preschool Q-set (PQ).
}
\references{
Baumrind, D. (1968). Manual for the Preschool Behaviour Q-set.
 Parental Research Project. Berkeley, CA: Institute of Human Development,
 University of California.

 Block, J. H., & Block, J. (1969). The California Child Q-Set.
 Berkeley, CA: Institute of Human Development, University of California.

 Posada, G., Waters, E., Crowell, J. A., & Lay, K. (1995).
 Is it easier to use a secure mother as a secure base?
 Attachment Q-Sort correlates of the adult attachment interview.
 Monographs of the Society for Research in Child Development, 60 (2/3), 133-145.

 Pederson, D. R., & Moran, G. (1995). A categorical description of infant-mother relationships
 in the home and its relation to Q-sort measures of infant-mother interaction.
 Monographs of the Society for Research in Child Development, 60 (2/3), 111-132.

 Pederson, D. R., Moran, G., & and Bento, S. (1999). Maternal Behaviour Q-sort (version 3.1).
 London, ON: Psychology Department, Western University.

 Waters, E. (1995). Appendix A: The attachment Q-set (Version 3. 0).
 Monographs of the Society for Research in Child Development, 60 (2/3), 234–246.

 Waters, E., Vaughn, B. E., Posada, G., & Kondo-Ikemura, K. (1995). Appendix D:
 Scoring key for Q-sort criteria and derived scales in the attachment Q-set
 and the attachment Q-set for infant macaques item sets.
 Monographs of the Society for Research in Child Development, 60 (2/3), 280-282.

 Waters, E., Noyes, D. M., Vaughn, B. E., & Ricks, M. (1985). Q-sort
 definitions of social competence and self-esteem: Discriminant validity of
 related constructs in theory and data. Developmental Psychology, 21, 508-522.
}
\keyword{datasets}
