% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsort_score.R
\name{qsort_score}
\alias{qsort_score}
\title{qsort_score}
\usage{
qsort_score(x, qset = names(qsets), item1, subj_id = NULL,
  group_id = NULL)
}
\arguments{
\item{x}{data frame. x should be in wide format, with each subject data in
one row and the scores for qset items in different columns. Columns with
qset item scores should be ordered sequentially from item 1 to the last
item.}

\item{qset}{"aqs" for Attachment Q-set (version 3.0) (Waters, 1995); "ccq" for
California Child Q-set (Block & Block, 1969); "mbqs" for Maternal
Behaviour Q-set (version 3.1) (Pederson et al., 1999); and "pq" for
Preschool Q-set (Baumrind, 1968 revised by Wanda Bronson).}

\item{item1}{column name of x containing item 1 qset score}

\item{subj_id}{optional. Column name of x with subjects' identification codes.}

\item{group_id}{optional. Column name of x with groups' identification codes.}
}
\value{
a data frame. This data frame will have a varying number of columns
 depending on the number of available criteria sorts and scales for the
 selected qset. Column names ending in "_c" refer to criteria sorts' scores, while
 column names ending in "_s" refer to scales' scores. The different types of criteria
 sorts and scales for each qset can be consulted in the documentation of
 qsets object (?qsets).
}
\description{
qsort_score returns a data frame with scores derived from criteria Q-sorts and
from scales created from subsets of items.
}
\examples{
data_ccq <- qsort_score(ex_qsort$ccq,
                       qset = "ccq",
                       item1 = "ccq1",
                       subj_id = "participant",
                       group_id = "classroom")
data_ccq

data_aqs <- qsort_score(ex_qsort$aqs,
                       qset = "aqs",
                       item1 = "aqs1")

}
\references{
Baumrind, D. (1968). Manual for the Preschool Behaviour Q-set.
 Parental Research Project. Berkeley, CA: Institute of Human Development,
 University of California.

 Block, J. H., & Block, J. (1969). The California Child Q-Set. Berkeley, CA:
 Institute of Human Development, University of California.

 Pederson, D. R., Moran, G., & and Bento, S. (1999). Maternal Behaviour
 Q-sort (version 3.1). London, ON: Psychology Department, Western University.

 Waters, E. (1995). Appendix A: The attachment Q-set (Version 3. 0).
 Monographs of the Society for Research in Child Development, 60, 234–246.
}
