% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsort_score.R
\name{qsort_score}
\alias{qsort_score}
\title{qsort_score}
\usage{
qsort_score(x, qset = c("pq", "ccq"), subj_id = NULL, group_id = NULL)
}
\arguments{
\item{x}{a data frame. x should be in wide format, with each subject data in
one row and the scores for CCQ and PQ items in different columns. CCQ and
PQ column names should be named sequentially from ccq1 to ccq100 and/or
from pq1 to pq72 (both CCQ and PQ Q-sort data can coexist in the same data
frame). qsort_score function is case insensitive for item column names
(i.e., CCQ1 or ccq1 will both work), but there should be no spaces in
column names. Q-set items should be scored 1 to 9 (most uncharacteristic to
most characteristic of the subject).}

\item{qset}{either qset = "pq" or qset = "ccq".}

\item{subj_id}{optional. Column name of x that includes subjects'
identification codes.}

\item{group_id}{optional. Column name of x that includes groups'
identification codes.}
}
\value{
a data frame. data frame will have 4 to 6 columns: subject id -
  (optional) subject identification; group id - (optional) group
  identification; scomp - social competence scores; sest - self-esteem
  scores; missing - number of missing item scores in x (if any); qset - q set
  identification.
}
\description{
qsort_score returns social competence and self-esteem scores (partial
correlations) for California Child Q-set (CCQ; Block and Block, 1969) or
Preschool Q-set data (PQ; Baumrind, 1968 revised by Wanda Bronson) using the
criterion scores presented in Waters et al. (1985), while controlling for
social desirability.
}
\examples{
data_ccq <- qsort_score(ex_qsort,
                        qset = "ccq",
                        subj_id = "participant",
                        group_id = "classroom")
data_ccq

}
\references{
Baumrind, D. (1968). Manual for the Preschool Behaviour Q-set.
  Parental Research Project. Berkeley, CA: Institute of Human Development,
  University of California.

  Block, J. H., & Block, J. (1969). The California Child Q-Set. Berkeley, CA:
  Institute of Human Development, University of California.

  Waters, E., Noyes, D. M., Vaughn, B. E., & Ricks, M. (1985). Q-sort
  definitions of social competence and self-esteem: Discriminant validity of
  related constructs in theory and data. Developmental Psychology, 21,
  508-522.
}
